[update] move Scope class to Model NS (#672)
